Refactor to remove oauth cruft
Removed unesseary branching for unused oauth code.
